requirements.txt
can
cantools
paho-mqtt
PyYAML